Add files via upload
Here is the code that I have written so far as next weeks homework. The task is to use the bisection function as a model to implement secant method to find the roots a function, specifically, the first 50 roots of the BesselJ0 function. I implemented this secant function. It appears that the The bisection was given to me and the Newton we wrote in class. I more used the Newton method as a model. Also attached is the lecture, so you know where I got the formula. I have two questions, how we figure out the optimized form of the formula for reduction of FLOP error/how that operation order reduces FLOP error. It appears that we were working on the Find Roots function during lecture but may have gotten cut short. The idea seems to be take two initial guesses of where a root is and repeatedly use a rootfinding method (i.e., Newton or Secant) to reduce the interval with each iteration? I could see how an array of guesses could be inputed into a function like this to give a table of zeros (per the questions request).
The second task is to do the same thing but with Newton's method, find first 50 zeros, and use the identity J0' = -J1 in the formula.